backend/requirements.txt
asgiref==3.2.10
django==2.2.16
django-cors-headers 
django-debug-toolbar==3.2.4
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields==3.2.1
gunicorn==20.0.4
pillow==9.0.0
psycopg2-binary==2.8.6
reportlab==3.6.3
sqlparse==0.3.1
